remove android app from folder

+++ apps/frappe/frappe/utils/backups.py
@@ -637,7 +637,7 @@ def delete_temp_backups(older_than=24):
 		file_list = os.listdir(get_backup_path())
 		for this_file in file_list:
 			this_file_path = os.path.join(get_backup_path(), this_file)
-			if is_file_old(this_file_path, older_than):
+			if os.path.isfile(this_file_path) and is_file_old(this_file_path, older_than):
 				os.remove(this_file_path)
